small updates
I noticed I had left out an if statement to determine if the user wanted to look up a specific Los Angeles team. Rather than improving the regex (which I had not learned about at the time) I maintained the style of code.
I also changed the winners function to checked the teams.txt file rather than the winners.txt file because the file was lost.